Added support for both python2 and python3 using six.

+++ scripts/lib/xpedite/profiler/environment.py
@@ -188,7 +188,7 @@ class Environment(object):
   def getCpuClockFrequencyRaw(self):
     """Returns the raw cpu clock frequency for localhost"""
     rawHz = self.getFullCpuInfo()['hz_advertised_raw'][0]
-    if self.tscHz / 10**9 != rawHz / 10**9:
+    if int(self.tscHz / 10**9) != int(rawHz / 10**9):
       LOGGER.error('Detected mismatch in estimated TSC (%s) vs raw frequency (%s)', self.tscHz, rawHz)
     return rawHz
